// src/sanity/types.ts
export type SocialLink = {
  _type: "socialLink";
  label?: string;
  url?: string;
};
// ...
export type SettingsQueryResult =
  | {
      title: null;
      description: null;
      nav: null;
      contactEmail: null;
      socialLinks: null;
      theme: null;
      indexPreviewMode: null;
    }
  | {
      title: string | null;
      description: null;
      nav: null;
      contactEmail: null;
      socialLinks: null;
      theme: null;
      indexPreviewMode: null;
    }
  | {
      title: string | null;
      description: string | null;
      nav: null;
      contactEmail: null;
      socialLinks: null;
      theme: null;
      indexPreviewMode: null;
    }
  | {
      title: string | null;
      description: string | null;
      nav: {
        home?: string;
        projects?: string;
        about?: string;
      } | null;
      contactEmail: string | null;
      socialLinks: Array<
        {
          _key: string;
        } & SocialLink
      > | null;
      theme: "dark" | "light" | null;
      indexPreviewMode: "cursor" | "fixed" | null;
    }
  | null;